test: extract css variable parser
