+++ tests/helpers/meditationContrast.test.js
@@ -1,26 +1,10 @@
 // @vitest-environment node
 import { describe, it, expect } from "vitest";
-import { readFileSync } from "fs";
-import { resolve } from "path";
 import { hex } from "wcag-contrast";
-import postcss from "postcss";
-
-function getBaseVars() {
-  const css = readFileSync(resolve("src/styles/base.css"), "utf8");
-  const vars = {};
-  const root = postcss.parse(css);
-  root.walkRules(":root", (rule) => {
-    rule.walkDecls((decl) => {
-      if (decl.prop.startsWith("--")) {
-        vars[decl.prop] = decl.value.trim();
-      }
-    });
-  });
-  return vars;
-}
+import { parseCssVars } from "./parseCssVars.js";
 
 describe("meditation quote color contrast", () => {
-  const vars = getBaseVars();
+  const vars = parseCssVars("src/styles/base.css");
 
   it("quote text vs block background is >= 4.5", () => {
     expect(vars["--color-text-inverted"]).toBeDefined();
